+++ src/Other/Helpers/JSON.php
@@ -47,6 +47,7 @@ class JSON
         $data = json_encode($value, $options, $depth);
 
         if ($data === false) {
+            var_dump($value);
             throw new JSONException(json_last_error());
         }